Match path predicate Assignment closes #12

+++ webflux_tutorial/webflux-demo/src/main/java/com/ziko/webfluxdemo/dto/Response.java
@@ -2,8 +2,8 @@ package com.ziko.webfluxdemo.dto;
 
 import java.util.Date;
 
-public record Response(Date date, int output) {
-    public Response(int output) {
+public record Response(Date date, double output) {
+    public Response(double output) {
         this(new Date(), output);
     }
